refactor edgecontainer API
