+++ pkg/controller/direct/edgecontainer/vpn_mappings.go
@@ -25,52 +25,6 @@ import (
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/controller/direct"
 )
 
-func EdgeContainerMachineObservedState_FromProto(mapCtx *direct.MapContext, in *pb.Machine) *krm.EdgeContainerMachineObservedState {
-	if in == nil {
-		return nil
-	}
-	out := &krm.EdgeContainerMachineObservedState{}
-	// MISSING: Name
-	out.CreateTime = direct.StringTimestamp_FromProto(mapCtx, in.GetCreateTime())
-	out.UpdateTime = direct.StringTimestamp_FromProto(mapCtx, in.GetUpdateTime())
-	out.HostedNode = direct.LazyPtr(in.GetHostedNode())
-	out.Version = direct.LazyPtr(in.GetVersion())
-	out.Disabled = direct.LazyPtr(in.GetDisabled())
-	return out
-}
-func EdgeContainerMachineObservedState_ToProto(mapCtx *direct.MapContext, in *krm.EdgeContainerMachineObservedState) *pb.Machine {
-	if in == nil {
-		return nil
-	}
-	out := &pb.Machine{}
-	// MISSING: Name
-	out.CreateTime = direct.StringTimestamp_ToProto(mapCtx, in.CreateTime)
-	out.UpdateTime = direct.StringTimestamp_ToProto(mapCtx, in.UpdateTime)
-	out.HostedNode = direct.ValueOf(in.HostedNode)
-	out.Version = direct.ValueOf(in.Version)
-	out.Disabled = direct.ValueOf(in.Disabled)
-	return out
-}
-func EdgeContainerMachineSpec_FromProto(mapCtx *direct.MapContext, in *pb.Machine) *krm.EdgeContainerMachineSpec {
-	if in == nil {
-		return nil
-	}
-	out := &krm.EdgeContainerMachineSpec{}
-	// MISSING: Name
-	out.Labels = in.Labels
-	out.Zone = direct.LazyPtr(in.GetZone())
-	return out
-}
-func EdgeContainerMachineSpec_ToProto(mapCtx *direct.MapContext, in *krm.EdgeContainerMachineSpec) *pb.Machine {
-	if in == nil {
-		return nil
-	}
-	out := &pb.Machine{}
-	// MISSING: Name
-	out.Labels = in.Labels
-	out.Zone = direct.ValueOf(in.Zone)
-	return out
-}
 func EdgeContainerVpnConnectionObservedState_FromProto(mapCtx *direct.MapContext, in *pb.VpnConnection) *krm.EdgeContainerVpnConnectionObservedState {
 	if in == nil {
 		return nil
